Fix Flask reloading when using SSE threads.
The default behavior of Flask is that it won't restart until all
threads finish, so we need a way to communicate to long running
SSE threads that they need to exit. Otherwise the Flask server
won't reload code, which makes development annoying and could lead
to complex problems in production. I think the solution here fixes
both, but I've only tested it in development.