MEM-205 Modify QuickPin to use Piscina instead of scraping Twitter
Add friends and followers from Twitter API into QuickPin. If a profile
is added that has friends/followers who are not in QuickPin, then a
"stub" profile is created: a profile that contains biographical data
but doesn't fetch posts, friends, followers, etc. In the future, we'll
add a feature to convert a stub into a full profile by fetching that
additional information.