MEM-205 Modify QuickPin to use Piscina instead of scraping Twitter
Create thumbnails of profile images. This allows for faster loading
of friends/followers list (get thumbnails instead of unresized original
image) and also gave me an opportunity to tweak the data models and
views, which is good because the avatar images had already gotten too
complex for me too handle.