Separate application DB account from admin DB account.
We have been running a single Postgres account for both administering
(building, dropping, migrating) the database as well as handling queries
from the application server. This approach turns out not to work well
with Agnostic migrations, so I added a separate pair of admin credentials
that can be used for building, dropping, and migrating. During installation,
the main application account will be given much lower privileges, which
is better for security anyway.

(Porting this feature from Avatar to QuickPin.)